move ctxmenu to page
