+++ src/routes/stream/[layout]/[[edit]]/+layout.svelte
@@ -1,11 +1,10 @@
 <script lang="ts">
     import { layoutNodes, type LayoutNodeUpdate } from '$lib/classes/layoutNodes.js'
     import { stateVariables, type StateVariableUpdate } from '$lib/classes/stateVariables.js'
-    import { activeNodeID, ctxMenu } from '$lib/stores/editor'
+    import { activeNodeID } from '$lib/stores/editor'
     import { supabase } from '$lib/supabaseClient.js'
 	import { wheel } from '$lib/stores/editor'
 
-    import ContextMenu from '$lib/components/menu/contextMenu.svelte'
     export let data
 
     layoutNodes.set(layoutNodes.getNodes(data.nodes))
@@ -28,15 +27,7 @@
             }
         }).subscribe()
 
-    const menu = {
-        "Save All": {disabled: true},
-        "Reset All": {disabled: true},
-        "Say hello": {action: () => alert("hello!")}
-    }
 </script>
-
-<svelte:window on:click={ctxMenu.close}/>
-<ContextMenu />
 
 
 {#if data.edit}
@@ -46,7 +37,7 @@
 
     <!-- svelte-ignore a11y-no-static-element-interactions -->
     <div id='scale-bg'
-        on:contextmenu|preventDefault={(e) => ctxMenu.open(e, menu)}
+        on:contextmenu|preventDefault
         on:wheel|preventDefault={wheel} 
         on:mousedown={() => activeNodeID.set(null)}/>
 {/if}
